dev-requirements.txt
nose
bumpversion
pyramid-debugtoolbar
pyramid-exclog